requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
cycler==0.11.0
EasyProcess==1.1
entrypoint2==1.1
filelock==3.9.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.3
idna==3.4
imageio==2.31.3
imutils==0.5.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.9.0
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.5.2
MouseInfo==0.1.3
mpmath==1.2.1
mss==9.0.1
multidict==6.0.4
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pafy==0.5.5
pandas==1.3.5
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytorch-lightning==2.1.2
pytweening==1.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sympy==1.11.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.8.30
torchmetrics==1.1.1
tqdm==4.66.1
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==2.0.4
Werkzeug==3.0.1
yarl==1.9.2
youtube-dl==2021.12.17
